Payment type

+++ scorex-transaction/src/main/scala/scorex/transaction/state/wallet/Payment.scala
@@ -3,19 +3,19 @@ package scorex.transaction.state.wallet
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{JsPath, Reads, Writes}
 
-case class Payment(amount: Double, fee: Double, sender: String, recipient: String)
+case class Payment(amount: Long, fee: Long, sender: String, recipient: String)
 
 object Payment {
   implicit val paymentWrites: Writes[Payment] = (
-    (JsPath \ "amount").write[Double] and
-      (JsPath \ "fee").write[Double] and
+    (JsPath \ "amount").write[Long] and
+      (JsPath \ "fee").write[Long] and
       (JsPath \ "sender").write[String] and
       (JsPath \ "recipient").write[String]
     ) (unlift(Payment.unapply))
 
   implicit val locationReads: Reads[Payment] = (
-    (JsPath \ "amount").read[Double] and
-      (JsPath \ "fee").read[Double] and
+    (JsPath \ "amount").read[Long] and
+      (JsPath \ "fee").read[Long] and
       (JsPath \ "sender").read[String] and
       (JsPath \ "recipient").read[String]
     ) (Payment.apply _)
